Move ViewModel tests to viewmodel package
